basic communication structure with exemplary packets

+++ backend/src/main/java/communication/CommunicationFactory.java
@@ -2,11 +2,9 @@ package communication;
 
 
 import login.LoginResponse;
-import login.RequestValidationResponse;
 import utils.Pair;
 
 import java.nio.file.Path;
-import java.util.function.Function;
 
 /**
  * Abstracts all functions of the communication between the java backend and the client.
@@ -25,10 +23,10 @@ public interface CommunicationFactory {
      * A function that the communication system can use to check if a login is valid
      * @param username - the username provided by the request
      * @param password - the password provided by the request
-     * @return - A pair consisting of a {@link LoginResponse}, a token, and the number of seconds until the token
-     * should expire.
+     * @return - A pair consisting of a {@link LoginResponse}, a token, and the unix timestamp in millis specifying
+     * when the token should expire.
      * If the {@link LoginResponse} is not Valid then the second argument will be null
      */
-    Pair<LoginResponse, Pair<String,Integer>> checkLogin(String username, String password);
+    Pair<LoginResponse, Pair<String,Long>> checkLogin(String username, String password);
 
 }
